move kafka logic into own file
