+++ lib/db.js
@@ -1,8 +1,8 @@
 import { PrismaClient } from '@prisma/client';
 import { ClickHouse } from 'clickhouse';
-import { Kafka } from 'kafkajs';
 import dateFormat from 'dateformat';
 import chalk from 'chalk';
+import { getKafkaService } from './kafka';
 import {
   MYSQL,
   MYSQL_DATE_FORMATS,
@@ -67,44 +67,15 @@ function getClickhouseClient() {
   });
 }
 
-function getKafkaClient() {
-  if (!process.env.KAFKA_URL) {
-    return null;
-  }
-
-  const url = new URL(process.env.KAFKA_URL);
-  const brokers = process.env.KAFKA_BROKER.split(',');
-
-  if (url.username.length === 0 && url.password.length === 0) {
-    return new Kafka({
-      clientId: 'umami',
-      brokers: brokers,
-    });
-  } else {
-    return new Kafka({
-      clientId: 'umami',
-      brokers: brokers,
-      ssl: true,
-      sasl: {
-        mechanism: 'plain',
-        username: url.username,
-        password: url.password,
-      },
-    });
-  }
-}
-
 const prisma = global.prisma || getPrismaClient(options);
 const clickhouse = global.clickhouse || getClickhouseClient();
-const kafka = global.kafka || getKafkaClient();
 
 if (process.env.NODE_ENV !== 'production') {
   global.prisma = prisma;
   global.clickhouse = clickhouse;
-  global.kafka = kafka;
 }
 
-export { prisma, clickhouse, kafka };
+export { prisma, clickhouse };
 
 export function getDatabase() {
   const type =
@@ -128,12 +99,6 @@ export function getAnalyticsDatabase() {
   if (!type) {
     return getDatabase();
   }
-
-  return type;
-}
-
-export function getKafkaService() {
-  const type = process.env.KAFKA_URL && process.env.KAFKA_URL.split(':')[0];
 
   return type;
 }
@@ -340,23 +305,4 @@ export async function runAnalyticsQuery(queries) {
     }
     return queries[CLICKHOUSE]();
   }
-}
-
-export async function kafkaProducer(params, topic) {
-  const producer = kafka.producer();
-  await producer.connect();
-
-  await producer.send({
-    topic,
-    messages: [
-      {
-        key: 'key',
-        value: JSON.stringify(params),
-      },
-    ],
-  });
-}
-
-export function getDateFormatKafka(date) {
-  return dateFormat(date, 'UTC:yyyy-mm-dd HH:MM:ss');
 }
